fix: split the pricing tests in three parts
