+++ .babelrc
@@ -1,4 +1,8 @@
 {
   "presets": ["next/babel"],
-  "plugins": ["istanbul"]
+  "env": {
+    "development": {
+      "plugins": ["istanbul"]
+    }
+  }
 }
